Changed structure of the defaults dirs.

+++ maps4fs/generator/map.py
@@ -72,6 +72,20 @@ class Map:
         self.custom_osm = custom_osm
         # endregion
 
+        # region custom DEM handling
+        self.custom_background_path: str | None = None
+        custom_dem = kwargs.get("custom_background_path", None)
+        if custom_dem and not os.path.isfile(custom_dem):
+            raise FileNotFoundError(f"Custom DEM file {custom_dem} does not exist.")
+
+        # Make a copy of the custom DEM to the map directory.
+        if custom_dem:
+            save_path = os.path.join(self.map_directory, os.path.basename(custom_dem))
+            shutil.copyfile(custom_dem, save_path)
+            self.logger.debug("Custom DEM copied to %s", save_path)
+            self.custom_background_path = save_path
+        # endregion
+
         # region main settings
         main_settings = MainSettings.from_map(self)
         main_settings_json = main_settings.to_json()
@@ -125,10 +139,6 @@ class Map:
 
         self.shared_settings = SharedSettings()
         self.components: list[Component] = []
-        custom_dem = kwargs.get("custom_background_path", None)
-        if custom_dem and not os.path.isfile(custom_dem):
-            raise FileNotFoundError(f"Custom DEM file {custom_dem} does not exist.")
-        self.custom_background_path = custom_dem
 
     def process_settings(self) -> None:
         """Checks the settings by predefined rules and updates them accordingly."""
